requirements.txt
Django==3.0.7
django-crispy-forms==1.9.0
gunicorn==20.0.4
stripe==2.45.0
Pillow==7.1.0
django-tables2==2.3.1
django-import-export==2.3.0
django-filter==2.3.0
django-bootstrap3==14.1.0
psycopg2-binary==2.8.5

